// src/types/mapbox-gl-draw.d.ts
  type DrawMode = DrawModes[keyof DrawModes];
// ...
  interface DrawActionableState {
    trash: boolean;
    combineFeatures: boolean;
    uncombineFeatures: boolean;
  }
// ...
  type DrawFeature =
    | DrawPoint
    | DrawLineString
    | DrawPolygon
    | DrawMultiFeature<"MultiPoint">
    | DrawMultiFeature<"MultiLineString">
    | DrawMultiFeature<"MultiPolygon">;

  interface MapMouseEvent extends MapboxMapMouseEvent {
    featureTarget: DrawFeature;
  }

  interface MapTouchEvent extends MapboxMapTouchEvent {
    featureTarget: DrawFeature;
  }
// ...
  interface DrawCustomModeThis {
    map: mapboxgl.Map;

    setSelected(features?: string | string[]): void;

    setSelectedCoordinates(
      coords: Array<{ coord_path: string; feature_id: string }>
    ): void;

    getSelected(): DrawFeature[];

    getSelectedIds(): string[];

    isSelected(id: string): boolean;

    getFeature(id: string): DrawFeature;

    select(id: string): void;

    delete(id: string): void;

    deleteFeature(id: string, opts?: any): void;

    addFeature(feature: DrawFeature): void;

    clearSelectedFeatures(): void;

    clearSelectedCoordinates(): void;

    setActionableState(actionableState: DrawActionableState): void;

    changeMode(mode: DrawMode, opts?: object, eventOpts?: object): void;

    updateUIClasses(opts: object): void;

    activateUIButton(name?: string): void;

    featuresAt(
      event: Event,
      bbox: BBox,
      bufferType: "click" | "tap"
    ): DrawFeature[];

    newFeature(geojson: GeoJSON): DrawFeature;

    isInstanceOf(type: string, feature: object): boolean;

    doRender(id: string): void;
  }

  interface DrawCustomMode<CustomModeState = any, CustomModeOptions = any> {
    onSetup?(
      this: DrawCustomModeThis & this,
      options: CustomModeOptions
    ): CustomModeState;

    onDrag?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapMouseEvent
    ): void;

    onClick?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapMouseEvent
    ): void;

    onMouseMove?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapMouseEvent
    ): void;

    onMouseDown?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapMouseEvent
    ): void;

    onMouseUp?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapMouseEvent
    ): void;

    onMouseOut?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapMouseEvent
    ): void;

    onKeyUp?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: KeyboardEvent
    ): void;

    onKeyDown?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: KeyboardEvent
    ): void;

    onTouchStart?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapTouchEvent
    ): void;

    onTouchMove?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapTouchEvent
    ): void;

    onTouchEnd?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapTouchEvent
    ): void;

    onTap?(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      e: MapTouchEvent
    ): void;

    onStop?(this: DrawCustomModeThis & this, state: CustomModeState): void;

    onTrash?(this: DrawCustomModeThis & this, state: CustomModeState): void;

    onCombineFeature?(
      this: DrawCustomModeThis & this,
      state: CustomModeState
    ): void;

    onUncombineFeature?(
      this: DrawCustomModeThis & this,
      state: CustomModeState
    ): void;

    toDisplayFeatures(
      this: DrawCustomModeThis & this,
      state: CustomModeState,
      geojson: GeoJSON,
      display: (geojson: GeoJSON) => void
    ): void;
  }